Move TaskCreateSimple to project/TaskCreateSimple/*
